Разработка третьего решения
