+++ hw05_parallel_execution/misc.go
@@ -1,0 +1,14 @@
+package hw05_parallel_execution //nolint:golint,stylecheck
+
+import (
+	"errors"
+	"sync"
+)
+
+var ErrErrorsLimitExceeded = errors.New("errors limit exceeded")
+
+type Errors struct {
+	count int
+	mx    sync.RWMutex
+}
+type Task func() error
